+++ src/sqlfmtpanel.h
@@ -4,6 +4,8 @@
 #include <QWidget>
 #include <QCheckBox>
 #include <QSpinBox>
+#include <QLabel>
+#include <QComboBox>
 #include <Qsci/qsciscintilla.h>
 
 class SqlFmtPanel : public QWidget {
@@ -19,8 +21,12 @@ private:
     QsciScintilla *m_output;
     QCheckBox *m_uppercase;
     QSpinBox *m_indent;
+    QComboBox *m_dialectCombo;
+    QLabel *m_statusLabel;
     QString m_inputText;
     void doFormat();
+    void setStatus(const QString &text, bool error = false);
+    void applySqlDialectKeywords();
 };
 
 #endif